+++ tests/Support/Stub/ExpressionBuilder.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Db\Tests\Support\Stub;
+
+use Yiisoft\Db\Expression\AbstractExpressionBuilder;
+
+class ExpressionBuilder extends AbstractExpressionBuilder
+{
+    protected function createSqlParser(string $sql): SqlParser
+    {
+        return new SqlParser($sql);
+    }
+}